check if user is deleted

+++ app/Console/Commands/ProfileDelete.php
@@ -93,7 +93,10 @@ class ProfileDelete extends Command
         if($this->confirm("Delete " . $profile->id . "?")){
             \App\SocialAccount::where('user_id',$profile->user_id)->delete();
             
-            $profile->user->delete();
+            if($profile->user){
+                $profile->user->delete();
+            }
+            
             $profile->delete();
         }
     }
